move function for converting filters to mongo style into helperfunctions file; use this in records table component
